requirements.txt
heapdict
matplotlib
numpy
numpydoc
pandas
pep8
Pillow
scipy
seaborn
